refactor controllers and add seeding data

+++ db/migrate/20221012074236_add_price_to_property.rb
@@ -1,5 +1,5 @@
 class AddPriceToProperty < ActiveRecord::Migration[6.1]
   def change
-    add_column :properties, :price, :integer
+    add_column :properties, :price, :float
   end
 end
